refactor / move safe_cast_to_index
